+++ acon-admin/src/lib/auth.ts
@@ -1,32 +1,7 @@
 // lib/auth.ts
 import { apiRequest, apiRequestFormData } from './api'
+import { getStoredCSRFToken, storeCSRFToken, clearStoredTokens } from '../utils/tokens'
 import { CSRFResponse, LoginCredentials, LoginResponse } from '../types/auth'
-
-/**
- * localStorage에서 CSRF 토큰 조회
- */
-export function getStoredCSRFToken(): string | null {
-  if (typeof window === 'undefined') return null
-  return localStorage.getItem('csrf_token')
-}
-
-/**
- * CSRF 토큰 저장
- */
-export function storeCSRFToken(token: string): void {
-  if (typeof window === 'undefined') return
-  localStorage.setItem('csrf_token', token)
-  console.log('🔑 CSRF token stored:', token.substring(0, 10) + '...')
-}
-
-/**
- * 토큰 삭제
- */
-export function clearStoredTokens(): void {
-  if (typeof window === 'undefined') return
-  localStorage.removeItem('csrf_token')
-  console.log('🗑️ All tokens cleared')
-}
 
 /**
  * CSRF 토큰 발급
@@ -106,7 +81,7 @@ export async function logoutUser(): Promise<void> {
       console.warn('⚠️ No CSRF token found, attempting logout without token')
       // CSRF 토큰이 없어도 로그아웃 시도 (세션 정리용)
       await apiRequestFormData<void>(
-        '/api/admin/logout',
+        '/admin/logout',
         new URLSearchParams(),
         {}
       )
@@ -129,3 +104,5 @@ export function isAuthenticated(): boolean {
   console.log('🔍 Authentication check:', hasToken ? 'Authenticated' : 'Not authenticated')
   return hasToken
 }
+
+export { getStoredCSRFToken, storeCSRFToken, clearStoredTokens } from '../utils/tokens'
