+++ acon-admin/src/types/ApiErrorResponse.ts
@@ -1,0 +1,4 @@
+export interface ApiErrorResponse {
+  code: number
+  message: string
+}
